chg: Split CLI into modules
