+++ remotepy/config.py
@@ -1,12 +1,17 @@
 import configparser
 import os
+from typing import Optional
 
 import typer
 import wasabi
 
+from remotepy.utils import get_instance_ids, get_instance_info, get_instances
+
 CONFIG_PATH = os.path.join(
     os.path.expanduser("~"), ".config", "remote.py/", "config.ini"
 )
+cfg = configparser.ConfigParser()
+cfg.read(CONFIG_PATH)
 
 app = typer.Typer()
 
@@ -63,44 +68,69 @@ def show(config_path: str = typer.Option(CONFIG_PATH, "--config", "-c")):
 
 @app.command()
 def add(
-    instance_name: str = typer.Argument(None, help="Instance name"),
+    instance_name: Optional[str] = typer.Argument(None),
     config_path: str = typer.Option(CONFIG_PATH, "--config", "-c"),
 ):
     """
-    Add a new default instance to the config file
+    Add a new default instance to the config file.
+
+    This function allows you to add a new default instance either by providing the name directly or
+    by selecting it from a list of currently running instances. Terminated instances are not included
+    in this list. If a name is directly provided, the function checks if an instance with this name
+    exists before setting it as the default.
+
+    Arguments:
+    config_path -- The path to the configuration file.
+    instance_name -- The name of the instance to add. This is an optional argument.
+                     If not provided, the function will prompt the user to select an instance from a list.
+
+    Returns:
+    None. But it modifies the configuration file with the new default instance.
     """
 
-    # Check whether the instance_name is already set in the config
+    if instance_name is None:
+        # No instance name provided. Fetch the list of currently running
+        # instances (excluding terminated ones)
 
-    if not instance_name:
+        instances = get_instances()
 
-        if cfg["DEFAULT"].get("instance_name"):
-            typer.secho(
-                f"Default instance already set as {cfg['DEFAULT']['instance_name']}",
-                fg=typer.colors.YELLOW,
-            )
+        # Get the instance ids for the instances
+        ids = get_instance_ids(instances)
 
-            # Ask if the user still wants to change the instance name
+        # Get other details like name, type etc for these instances
+        names, _, _, instance_types, _ = get_instance_info(instances)
 
-            add_new = typer.confirm("Do you want to add a new default instance?")
+        # Prepare a formatted string to display these instance details as a
+        # table
+        header = ["Number", "Name", "InstanceId", "Type"]
+        aligns = ["l", "l", "l", "l"]
+        data = [
+            (i, name, id, it)
+            for i, (name, id, it) in enumerate(zip(names, ids, instance_types), 1)
+        ]
 
-            if add_new:
-                instance_name = typer.prompt("Instance name")
+        # Print the instances table
+        formatted = wasabi.table(data, header=header, divider=True, aligns=aligns)
+        typer.secho(formatted, fg=typer.colors.YELLOW)
 
-        # If not default exists in the config, then prompt for instance name
+        # Prompt the user to select an instance from the table
+        instance_number = typer.prompt("Select a instance by number", type=int)
 
+        # Validate the user input
+
+        if 1 <= instance_number <= len(names):
+            # If the input is valid, set the instance name to the selected one
+            instance_name = names[instance_number - 1]
         else:
-            typer.secho("No default instance found", fg=typer.colors.YELLOW)
-            instance_name = typer.prompt("Set the default instance name")
+            # Invalid input. Display an error message and exit.
+            typer.secho("Invalid number. No changes made", fg=typer.colors.YELLOW)
 
-    # If the instance_name has been set, then add it to the config
+            return
 
-    if instance_name:
-        cfg.set("DEFAULT", "instance_name", instance_name)
-        write_config(cfg, config_path)
-        typer.secho(f"Default instance set to {instance_name}", fg=typer.colors.GREEN)
-    else:
-        typer.secho("No changes made", fg=typer.colors.YELLOW)
+    # If an instance name was directly provided or selected from the list, update the configuration file
+    cfg.set("DEFAULT", "instance_name", instance_name)
+    write_config(cfg, config_path)
+    typer.secho(f"Default instance set to {instance_name}", fg=typer.colors.GREEN)
 
 
 if __name__ == "__main__":
